KRELL-68: Asset & lib loads lost after recompile
Write nses with requires to an "append-only" cache. After the
build completes we use this set of nses to load the analysis to
compute assets and requires. This always works because the analysis
cache will always have up-to-date information. We cannot rely
on the compiler itself because the compiler will just avoid doing any
work if it can - that is it will never load the analysis for files
it never needs to look at.

Fixes #68