KRELL-100: Remove the Android optimization
The loaded check was problematic - it naively checked to see if the
namespace had already been provided for. However in some cases like
`goog.debug` namespaces which are beneath that `goog.debug.Foo` may
not actually depend on `goog.debug` and load before. This would of
course make the existence return the wrong answer.

Just remove this logic.