Drop isProvided hack (#117)
* comment out isProvided hack

* use ClojureScript version with idempotent boostrap/install-repl-goog-require

always establish printing and REPL patches when connecting to environment even if GCL base & cljs.core are present

* fixup logging

* always monkeypatch goog.isProvided_ and goog.require *after* the main namespace is loaded.
this ensures goog.modules load properly. this code is now moved out of the server and
into the client code where we can be certain it will occur after a React Native refresh.

always establish printing after the main namespace is loaded. This addresses the sporadic
printing issues where the output doesn't print to the REPL.